List Hatası Düzeltildi
