--- bin/ClassMetotDemo.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.2.32630.192
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ClassMetotDemo", "ClassMetotDemo.csproj", "{08C9B8A1-B94D-4062-B52E-75CFF55F59BD}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{08C9B8A1-B94D-4062-B52E-75CFF55F59BD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{08C9B8A1-B94D-4062-B52E-75CFF55F59BD}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{08C9B8A1-B94D-4062-B52E-75CFF55F59BD}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{08C9B8A1-B94D-4062-B52E-75CFF55F59BD}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {170EA09C-DEC6-4BD9-8106-6E022BD0C48C}
-	EndGlobalSection
-EndGlobal
