--- bin/Musteri.cs
@@ -1,21 +1,0 @@
-﻿using System;
-
-namespace ClassMetotDemo
-{
-    public class Musteri
-    {
-     
-        public int Id { get; set; }
-        public string Name { get; set; }
-        public string SurName { get; set; }
-        public char Gender { get; set; }
-        public DateTime BirthDate { get; set; }
-        public string MaritalStatus { get; set; }
-        public string City { get; set; }
-        public string Country { get; set; }
-        public string Phone { get; set; }
-        public string Email { get; set; }
-        
-        
-    }
-}
