--- bin/Program.cs
@@ -1,13 +1,0 @@
-﻿
-
-namespace ClassMetotDemo
-{
-    class Program
-    {
-        static void Main()
-        {
-            Menu menu = new Menu();
-            menu.Menus();
-        }
-    }
-}
