--- bin/bin/Debug/netcoreapp3.1/ClassMetotDemo.deps.json
@@ -1,23 +1,0 @@
-{
-  "runtimeTarget": {
-    "name": ".NETCoreApp,Version=v3.1",
-    "signature": ""
-  },
-  "compilationOptions": {},
-  "targets": {
-    ".NETCoreApp,Version=v3.1": {
-      "ClassMetotDemo/1.0.0": {
-        "runtime": {
-          "ClassMetotDemo.dll": {}
-        }
-      }
-    }
-  },
-  "libraries": {
-    "ClassMetotDemo/1.0.0": {
-      "type": "project",
-      "serviceable": false,
-      "sha512": ""
-    }
-  }
-}
