--- bin/bin/Debug/netcoreapp3.1/ClassMetotDemo.runtimeconfig.dev.json
@@ -1,9 +1,0 @@
-{
-  "runtimeOptions": {
-    "additionalProbingPaths": [
-      "C:\\Users\\Turan\\.dotnet\\store\\|arch|\\|tfm|",
-      "C:\\Users\\Turan\\.nuget\\packages",
-      "C:\\Program Files (x86)\\Microsoft Visual Studio\\Shared\\NuGetPackages"
-    ]
-  }
-}
